Separated components for prev/next and number buttons
