+++ react-multiple-components/my-app/src/NextPrevButton.tsx
@@ -1,0 +1,8 @@
+type Props = {
+  text: string;
+  onClick: () => void;
+};
+
+export function PrevNextButtons({ text, onClick }: Props) {
+  return <button onClick={onClick}>{text}</button>;
+}
